wider scope .pdb atom lines
As by [1] as by experience writing .pdb with openbabel [2],
lines about atom coordinates may start by either 'ATOM', or
'HETATOM' as keyword.  The addition of the later widens the
scope of calculate_rmsd.py, an additional condition to work
well with e.g. openbabel-written .pdb.  The idea however was
deposit earlier by 'nabraham47' as issue #56 (2019-May-30)
to the general attention.
[1] http://www.wwpdb.org/documentation/file-format-content/format33/sect9.html#ATOM
[2] www.openbabel.org