Bug fix, capitalization of elements in .xyz files
In previous commits, calculate_rmsd.py wrote .xyz files where all
characters of the element where capitalized; e.g., not only 'C' for
carbon, but 'CL' or 'BR' for chlorine and bromine, instead of 'Cl'
or 'Br'.  This however may hinder further processing of the data.
Open Babel (version 3.1.0) for example replaces an element entry
like 'BR' in an .xyz file to be exported as either .xyz, or .pdb
just by '*'.

This bugfix addresses the issue for .xyz files to be written by a
modified calculate_rmsd.py.  The revised function is monitored by
an additional automated test in tests.py and reference input files.